Extract config helper method in to its own file
